Bug: incorrect where clause between issue & attachment
Fixes #67

On add a comment:
A partial record created for an uploaded file that includes the file details without
the FKs for comment or issue. On saving the comment or creating a new issue, these
FKs are populted. However, if the comment are not saved, this creates orphan files
and database records!